+++ .gitignore
@@ -148,6 +148,7 @@ dmypy.json
 
 # pytype static type analyzer
 .pytype/
+.vscode/
 
 # Cython debug symbols
 cython_debug/
