+++ README.md
@@ -1,1 +1,39 @@
 # batch-find-replace
+This script will search and replace a charater pattern in a batch of files located in a specified folder.
+
+## Requirements
+ - python 3.9 installed on your system
+
+## Usage
+To use this script clone this repo to your local machine. Open the folder containing the main.py file in your terminal and run:
+
+    $ python main.py
+
+
+You will be prompted to add a path to the folder containing the files. Next, add a character pattern to be replaced, and  the character to replace it. 
+
+## Author
+
+👤 **Hubert MUGABO**
+
+- Github: [@mugberto](https://github.com/mugberto)
+- Twitter: [@mugberto](https://twitter.com/mugberto)
+- Linkedin: [mugberto](https://www.linkedin.com/in/hubert-mugabo-23144b6a/)
+
+## 🤝 Contributing
+
+Contributions, issues and feature requests are welcome!
+
+Feel free to check the [issues page](https://github.com/mugberto/batch-find-replace/issues).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Microverse
+
+## 📝 License
+
+This project is [MIT Lisenced](/LICENSE)
